Conrad's plotly
Conrad's plotly implemented in the shiny server - works, but the UCL are not actual ones (need to find a way to store all UCLs).
Plot works, but need to reorganize the data behind, then review graph codes.

Only one syndrome at a time - need to have a different page with a syndromic summary